ai-service/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
PyQt5>=5.15.0
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.10.0
filterpy>=1.4.5
pandas>=2.0.0
pillow>=9.5.0